+++ NEWS.md
@@ -8,6 +8,7 @@ WIP
 * Added "query = list(...)" to the created function if the original URL
   had query parameters
 * Nicer formatting of make_req() functions via styler
+* Added option `curlconverter.quiet` to disable all stdout output
 
 ## 0.8.0
 * support for `-v / --verbose` and `-u / --user` `curl` command-line parameters.